Se reorganizaron los metodos de la clase Carrera

+++ src/main/java/dominio/Dado.java
@@ -5,10 +5,10 @@ package dominio;
 public class Dado {
     
     
-    public int lanzarDado () {
+    public double lanzarDado () {
         
-        int resultadoDado=(int) Math.floor(Math.random()*6+1);
-        int avanceMetros= resultadoDado*100;
+        double resultadoDado=Math.floor(Math.random()*6+1);
+        double avanceMetros= resultadoDado*100;
         return avanceMetros;
     }
